msme_chatbot_agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "msme-chatbot-agents"
version = "0.1.0"
description = "MSME Chatbot Agents with LangGraph Swarm and MCP Integration"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "MSME Chatbot Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # AI/ML Frameworks
    "langchain==0.3.26",
    "langchain-anthropic==0.3.9",
    "langchain-community==0.3.26",
    "langchain-core>=0.3.68,<0.3.77",
    "langchain-experimental==0.3.4",
    "langchain-google-genai==2.1.12",
    "langchain-mcp-adapters==0.1.7",
    "langchain-mongodb==0.6.2",
    "langchain-openai>=0.3.27,<0.3.29",
    "langchain-text-splitters==0.3.8",
    "langchain-qdrant>=0.2.1",
    "langgraph==0.5.4",
    "langgraph-checkpoint==2.1.1",
    "langgraph-checkpoint-mongodb==0.1.4",
    "langgraph-prebuilt==0.5.2",
    "langgraph-sdk>=0.1.72,<0.1.75",
    "langgraph-supervisor==0.0.27",
    "langgraph-swarm==0.0.13",
    "langsmith>=0.4.5,<0.4.9",

    # OpenAI & Anthropic
    "openai>=1.95.1,<1.98",
    "anthropic==0.49.0",

    # Database
    "pymongo==4.11.3",
    "motor==3.7.0",
    "sqlalchemy==2.0.38",
    "psycopg2-binary==2.9.10",
    "qdrant-client>=1.15.1",

    # Data Processing
    "pandas==2.2.3",
    "numpy==2.2.3",
    "pandasql==0.7.3",
    "pyarrow==19.0.1",

    # MCP Tools
    "mcp>=1.10.1,<1.17",
    "mcp-server==0.1.4",
    "cli-mcp==0.18.0",

    # Utilities
    "python-dotenv==1.1.1",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.8.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "aiohttp==3.11.13",
    "requests==2.32.4",
    "tenacity==9.1.2",
    "click==8.1.8",
    "rich==14.0.0",

    # Development Tools
    "ipython==9.0.1",
    "pyyaml==6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.wheel.force-include]
"__init__.py" = "msme_chatbot_agents/__init__.py"
"mcp_agents.py" = "msme_chatbot_agents/mcp_agents.py"
"mcp_agent_tools.py" = "msme_chatbot_agents/mcp_agent_tools.py"

[tool.hatch.build.targets.sdist]
include = [
    "*.py",
    "README.md",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.black]
line-length = 120
target-version = ["py312"]
